requirements.txt
chainlit==2.0.0
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-text-splitters==0.2.4
langchain-groq==0.1.9
chromadb==0.5.4
sentence-transformers==2.7.0
transformers==4.41.2
huggingface-hub==0.23.4
pypdf2==3.0.1
python-dotenv==1.0.1
tiktoken==0.7.0